feat!: Implement `Extend` for `BitSet`
For that to happen without nightly compiler we had to implement `new` for every primitive integer type. That also means that for generic context over primitive interger will fail to work, that means that this change is a API breaking change. For the same reason, a new method called `with_default` was added.

BREAKING CHANGE: API change